+++ email/src/email/message/config.rs
@@ -1,5 +1,7 @@
 use serde::{Deserialize, Serialize};
 
+#[cfg(feature = "account-sync")]
+use super::sync::config::MessageSyncConfig;
 use super::{
     add::config::MessageWriteConfig, get::config::MessageReadConfig,
     send::config::MessageSendConfig,
@@ -16,4 +18,8 @@ pub struct MessageConfig {
 
     /// Configuration dedicated to message sending.
     pub send: Option<MessageSendConfig>,
+
+    #[cfg(feature = "account-sync")]
+    /// Configuration dedicated to message sending.
+    pub sync: Option<MessageSyncConfig>,
 }