+++ email/src/sync/mod.rs
@@ -8,7 +8,6 @@ pub mod pool;
 pub mod report;
 
 use advisory_lock::{AdvisoryFileLock, FileLockError, FileLockMode};
-use chrono::{DateTime, Local};
 use dirs::cache_dir;
 use log::{debug, trace};
 use std::{
@@ -29,9 +28,9 @@ use crate::{
     folder::{
         self,
         sync::{
+            config::FolderSyncStrategy,
             hunk::{FolderName, FolderSyncHunk},
             patch::FolderSyncPatch,
-            FolderSyncStrategy,
         },
     },
     maildir::{config::MaildirConfig, MaildirContextBuilder},
@@ -63,7 +62,7 @@ pub struct SyncBuilder<L: BackendContextBuilder, R: BackendContextBuilder> {
     cache_dir: Option<PathBuf>,
     handler: Option<Arc<SyncEventHandler>>,
     dry_run: Option<bool>,
-    filters: Option<SyncFilters>,
+    folder_filter: Option<FolderSyncStrategy>,
 }
 
 impl<L: BackendContextBuilder + 'static, R: BackendContextBuilder + 'static> SyncBuilder<L, R> {
@@ -80,7 +79,7 @@ impl<L: BackendContextBuilder + 'static, R: BackendContextBuilder + 'static> Syn
             cache_dir: None,
             handler: None,
             dry_run: None,
-            filters: None,
+            folder_filter: Default::default(),
         }
     }
 
@@ -178,17 +177,8 @@ impl<L: BackendContextBuilder + 'static, R: BackendContextBuilder + 'static> Syn
         self.dry_run.unwrap_or_default()
     }
 
-    pub fn set_some_folders_filter(&mut self, folders: Option<impl Into<FolderSyncStrategy>>) {
-        let folders = folders.map(Into::into);
-        match self.filters.as_mut() {
-            Some(filters) => filters.folders = folders,
-            None => {
-                self.filters = Some(SyncFilters {
-                    folders,
-                    ..Default::default()
-                })
-            }
-        }
+    pub fn set_some_folders_filter(&mut self, filter: Option<impl Into<FolderSyncStrategy>>) {
+        self.folder_filter = filter.map(Into::into);
     }
 
     pub fn set_folders_filter(&mut self, folders: impl Into<FolderSyncStrategy>) {
@@ -206,100 +196,6 @@ impl<L: BackendContextBuilder + 'static, R: BackendContextBuilder + 'static> Syn
     pub fn with_folders_filter(mut self, folders: impl Into<FolderSyncStrategy>) -> Self {
         self.set_folders_filter(folders);
         self
-    }
-
-    pub fn get_folders_filter(&self) -> FolderSyncStrategy {
-        self.filters
-            .as_ref()
-            .and_then(|f| f.folders.as_ref())
-            .cloned()
-            .unwrap_or_default()
-    }
-
-    pub fn set_some_from_date_filter(&mut self, date: Option<impl Into<DateTime<Local>>>) {
-        let date = date.map(Into::into);
-        match self.filters.as_mut() {
-            Some(filters) => match filters.date_range.as_mut() {
-                Some(date_range) => date_range.set_some_from(date),
-                None => {
-                    filters.date_range = Some(SyncDateRangeFilter::default().with_some_from(date))
-                }
-            },
-            None => {
-                self.filters = Some(SyncFilters {
-                    date_range: Some(SyncDateRangeFilter::default().with_some_from(date)),
-                    ..Default::default()
-                })
-            }
-        }
-    }
-
-    pub fn set_from_date_filter(&mut self, date: impl Into<DateTime<Local>>) {
-        self.set_some_from_date_filter(Some(date));
-    }
-
-    pub fn with_some_from_date_filter(mut self, date: Option<impl Into<DateTime<Local>>>) -> Self {
-        self.set_some_from_date_filter(date);
-        self
-    }
-
-    pub fn with_from_date_filter(mut self, date: impl Into<DateTime<Local>>) -> Self {
-        self.set_from_date_filter(date);
-        self
-    }
-
-    pub fn find_from_date_filter(&self) -> Option<&DateTime<Local>> {
-        self.filters
-            .as_ref()
-            .and_then(|f| f.date_range.as_ref())
-            .and_then(|r| r.find_from())
-    }
-
-    pub fn get_date_range_filter(&self) -> SyncDateRangeFilter {
-        self.filters
-            .as_ref()
-            .and_then(|f| f.date_range.as_ref())
-            .cloned()
-            .unwrap_or_default()
-    }
-
-    pub fn set_some_to_date_filter(&mut self, date: Option<impl Into<DateTime<Local>>>) {
-        let date = date.map(Into::into);
-        match self.filters.as_mut() {
-            Some(filters) => match filters.date_range.as_mut() {
-                Some(date_range) => date_range.set_some_to(date),
-                None => {
-                    filters.date_range = Some(SyncDateRangeFilter::default().with_some_to(date))
-                }
-            },
-            None => {
-                self.filters = Some(SyncFilters {
-                    date_range: Some(SyncDateRangeFilter::default().with_some_to(date)),
-                    ..Default::default()
-                })
-            }
-        }
-    }
-
-    pub fn set_to_date_filter(&mut self, date: impl Into<DateTime<Local>>) {
-        self.set_some_to_date_filter(Some(date));
-    }
-
-    pub fn with_some_to_date_filter(mut self, date: Option<impl Into<DateTime<Local>>>) -> Self {
-        self.set_some_to_date_filter(date);
-        self
-    }
-
-    pub fn with_to_date_filter(mut self, date: impl Into<DateTime<Local>>) -> Self {
-        self.set_to_date_filter(date);
-        self
-    }
-
-    pub fn find_to_date_filter(&self) -> Option<&DateTime<Local>> {
-        self.filters
-            .as_ref()
-            .and_then(|f| f.date_range.as_ref())
-            .and_then(|r| r.find_to())
     }
 
     pub fn get_left_cache_builder(&self) -> Result<BackendBuilder<MaildirContextBuilder>> {
@@ -343,8 +239,7 @@ impl<L: BackendContextBuilder + 'static, R: BackendContextBuilder + 'static> Syn
                 self.right_builder.clone(),
                 self.handler.clone(),
                 self.get_dry_run(),
-                self.get_folders_filter(),
-                self.get_date_range_filter(),
+                self.folder_filter,
             )
             .await?,
         );
@@ -479,89 +374,6 @@ impl fmt::Display for SyncDestination {
         match self {
             Self::Left => write!(f, "left"),
             Self::Right => write!(f, "right"),
-        }
-    }
-}
-
-/// The synchronization filters.
-#[derive(Clone, Debug, Default, Eq, PartialEq)]
-pub struct SyncFilters {
-    /// Filter folders using the given strategy.
-    folders: Option<FolderSyncStrategy>,
-
-    /// Filter envelopes using the given date range.
-    ///
-    /// This filter operates on the `Date` header of the envelopes.
-    date_range: Option<SyncDateRangeFilter>,
-}
-
-/// The date range synchronization filter.
-#[derive(Clone, Debug, Default, Eq, PartialEq)]
-pub struct SyncDateRangeFilter {
-    from: Option<DateTime<Local>>,
-    to: Option<DateTime<Local>>,
-}
-
-impl SyncDateRangeFilter {
-    pub fn set_some_from(&mut self, date: Option<impl Into<DateTime<Local>>>) {
-        self.from = date.map(Into::into);
-    }
-
-    pub fn set_from(&mut self, date: impl Into<DateTime<Local>>) {
-        self.set_some_from(Some(date));
-    }
-
-    pub fn with_some_from(mut self, date: Option<impl Into<DateTime<Local>>>) -> Self {
-        self.set_some_from(date);
-        self
-    }
-
-    pub fn with_from(mut self, date: impl Into<DateTime<Local>>) -> Self {
-        self.set_from(date);
-        self
-    }
-
-    pub fn find_from(&self) -> Option<&DateTime<Local>> {
-        self.from.as_ref()
-    }
-
-    pub fn set_some_to(&mut self, date: Option<impl Into<DateTime<Local>>>) {
-        self.to = date.map(Into::into);
-    }
-
-    pub fn set_to(&mut self, date: impl Into<DateTime<Local>>) {
-        self.set_some_to(Some(date));
-    }
-
-    pub fn with_some_to(mut self, date: Option<impl Into<DateTime<Local>>>) -> Self {
-        self.set_some_to(date);
-        self
-    }
-
-    pub fn with_to(mut self, date: impl Into<DateTime<Local>>) -> Self {
-        self.set_to(date);
-        self
-    }
-
-    pub fn find_to(&self) -> Option<&DateTime<Local>> {
-        self.to.as_ref()
-    }
-
-    pub fn matches(&self, date: &DateTime<Local>) -> bool {
-        self.matches_from(date) && self.matches_to(date)
-    }
-
-    pub fn matches_from(&self, date: &DateTime<Local>) -> bool {
-        match self.from.as_ref() {
-            Some(from) => from >= date,
-            None => true,
-        }
-    }
-
-    pub fn matches_to(&self, date: &DateTime<Local>) -> bool {
-        match self.to.as_ref() {
-            Some(to) => to <= date,
-            None => true,
         }
     }
 }